+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add success and failure icons for different resolutions
+
+### Fixed
+
+- Fix icons for Light and Dark mode
+
 ## [1.2.0] - 2019-10-29
 
 ### Added